pyproject.toml
[project]
name = "swiss-life"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "baml-cli>=0.1.0",
    "baml-py==0.203.1",
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.14.1",
    "uvicorn>=0.35.0",
]
